Change _buildKey method to use more "standard" key approach.
The keys needed to be aware of global groups and multisite. Unfortunately, in an attempt to use the native namespacing feature with the memcached PECL extension, these were overlooked. Thanks for Scott Taylor and Andrew Nacin for the bug report and Ryan Boren and Sivel for the borrowed "correct/standard" method.