Do not store "false" when getting a "no_mc_group" value that does not exist.
The "get" methods check if the the value is available in the internal cache array. If it isn't and it is a no_mc_group value, it would set the value to false. It then sets that to the internal cache value. Therefore, if it tries to get this value and fails, it actually ends up setting the value, which is unexpected behavior. This patch fixes that. I will likely continue to improve this.