Instead of checking false, getResultCode is checked in the "get" methods.
Because a fetched value could be (bool) false, checking it against strict (bool) false may result in a false value not being stored in the internal cache. To fix this, the result code returned from Memcached is checked, which will indicate whether the key was found or not. This allows for differentiating between a value of false and a key that was not located.

Additionally, rather than setting the $value to false when getting from a no_mc_group, the value is immediately returned as false.