Better detection of cache_cb and cas_token in "get" methods.
Previously had difficulties finding a good way to route requests to Memcached if user sent an undefined variable to the "get" methods as the 4th variable. This commit adds checks to see if the 3rd or 4th argument is explicitly sent. If it is the request is routed to Memcached, unless the group is a no_mc_group. In that case, the 3rd and 4th arguments are ignored and the request is routed to the internal cache.