Fixed append behavior and properly handled calling get from wrapper.
It turns out that when appending values, memcached will change the type of the appended value to that of the initial value. In order to make sure the internal cache and memcached return the same values, a routine for proper casting was implemented.

The wp_cache_get method was always prompting the cache to get the value from memcached. This was because the class' method thought there were more args sent to the function than there really were. This was fixed by checking the number of args sent to wp_cache_get.