Added a method for handling prepends and appends.
In order to combine values in prepend and append operations, memcached first combines the values, then cast the combined value to the type of the original value. It is important that this behavior is mimicked in the internal cache so that values returned from memcached and the internal cache are identical. A new method was added to handle this procedure. Additionally, documentation was updated to reflect the confusion that this can lead to. Note: addresses issue #18.